fix: rubocop violation by splitting instrumentation to helper
