+++ lib/integrations/llm_instrumentation.rb
@@ -2,9 +2,11 @@
 
 require 'opentelemetry_config'
 require_relative 'llm_instrumentation_constants'
+require_relative 'llm_instrumentation_helpers'
 
 module Integrations::LlmInstrumentation
   include Integrations::LlmInstrumentationConstants
+  include Integrations::LlmInstrumentationHelpers
 
   def tracer
     @tracer ||= OpentelemetryConfig.tracer
@@ -107,38 +109,5 @@ module Integrations::LlmInstrumentation
     params[:metadata].each do |key, value|
       span.set_attribute(format(ATTR_LANGFUSE_METADATA, key), value.to_s)
     end
-  end
-
-  def set_completion_attributes(span, result)
-    set_completion_message(span, result)
-    set_usage_metrics(span, result)
-    set_error_attributes(span, result)
-  end
-
-  def set_completion_message(span, result)
-    message = result[:message] || result.dig('choices', 0, 'message', 'content')
-    return if message.blank?
-
-    span.set_attribute(ATTR_GEN_AI_COMPLETION_ROLE, 'assistant')
-    span.set_attribute(ATTR_GEN_AI_COMPLETION_CONTENT, message)
-  end
-
-  def set_usage_metrics(span, result)
-    usage = result[:usage] || result['usage']
-    return if usage.blank?
-
-    span.set_attribute(ATTR_GEN_AI_USAGE_INPUT_TOKENS, usage['prompt_tokens']) if usage['prompt_tokens']
-    span.set_attribute(ATTR_GEN_AI_USAGE_OUTPUT_TOKENS, usage['completion_tokens']) if usage['completion_tokens']
-    span.set_attribute(ATTR_GEN_AI_USAGE_TOTAL_TOKENS, usage['total_tokens']) if usage['total_tokens']
-  end
-
-  def set_error_attributes(span, result)
-    error = result[:error] || result['error']
-    return if error.blank?
-
-    error_code = result[:error_code] || result['error_code']
-    span.set_attribute(ATTR_GEN_AI_RESPONSE_ERROR, error.to_json)
-    span.set_attribute(ATTR_GEN_AI_RESPONSE_ERROR_CODE, error_code) if error_code
-    span.status = OpenTelemetry::Trace::Status.error("API Error: #{error_code}")
   end
 end
